==> 1.2_corpus_levenshtein_analysis.ipynb <==
import pandas as pd
from rapidfuzz.distance import Levenshtein
from tqdm import tqdm
import csv

# Loading SSCORPUS with similarity scores into a DataFrame

column_names = ['Professional_Sentence', 'Simple_Sentence', 'Similarity_Score']
df_sscorpus = pd.read_csv("sscorpus", sep='\t', header=None, names=column_names)
#df_sscorpus.head(100)

# Calculating the Levenshtein scores for the simple and professional sentences

sscorpus_levenshtein_weighted = []
for index, row in tqdm(df_sscorpus.iterrows()):
    professional = row['Professional_Sentence']
    simple = row['Simple_Sentence']
    normalized_distance = Levenshtein.distance(simple, professional, weights=(3,1,2)) / max(len(simple), len(professional))
    sscorpus_levenshtein_weighted.append([simple, professional, round(normalized_distance, 2)])

# Converting list into pandas Dataframe for easier analysis

column_names_levenshtein = ['Professional_Sentence', 'Simple_Sentence', 'Levenshtein_Score']
df_sscorpus_levenshtein = pd.DataFrame (sscorpus_levenshtein_weighted, columns = column_names_levenshtein)
#df_sscorpus_levenshtein.head(100)

# Analysis for determining the number of entries lower or equal than a score

# set the score according to need
score = 3.0
df_sscorpus_levenshtein.loc[ (df_sscorpus_levenshtein['Levenshtein_Score'] <= score) ].shape[0]

# Analysis for determining the number of entries in a range

# set the min_score and max_score according to need
min_score = 2.8
max_score = 3.0

df_sscorpus_levenshtein.loc[ (df_sscorpus_levenshtein['Levenshtein_Score'] > min_score) & (df_sscorpus_levenshtein['Levenshtein_Score'] <= max_score) ].shape[0]
