==> 2.1_preprocess_corpus_multi_label_true.ipynb <==
import pandas as pd
import csv
import pickle

# Load the SSCORPUS that is preprocessed with Levenshtein

columns = ['Simple', 'Professional', 'Levenshtein']
path = "1_sscorpus_levenshtein_calculated/sscorpus_levenshtein_final.csv"
df_levenshtein = pd.read_csv(path, sep=',', header=None, names=columns)

simple_list = df_levenshtein['Simple'].values.tolist()
professional_list = df_levenshtein['Professional'].values.tolist()

# Load the categorized SSCORPUS

dump_root_path = "2_sscorpus_categorized/multi_label_true/"

fin = open(dump_root_path+"scorpus_categorized_raw", 'rb')
results = pickle.load(fin)
fin.close()

# Preprocess and save the categorized SSCORPUS for further annotation for only specific categories

min_score = 0.3
categorized_sscorpus_specific = []
for index_result, result in enumerate(results):
    for index_score, score in enumerate(result['scores']):
        if round(score, 2) >= min_score:
            categorized_sscorpus_specific.append([ simple_list[index_result], result['sequence'], result['labels'][index_score], round(score, 2) ])
            
with open(dump_root_path+'initial/sscorpus_categorized_preprocessed_specific.csv', 'w+', newline='') as f:
    write = csv.writer(f)
    write.writerows(categorized_sscorpus_specific)
    f.close()

# Preprocess and save the categorized SSCORPUS for further annotation for the full corpus
min_score = 0.3
categorized_sscorpus_full = []
for index, result in enumerate(results):
    if round(result['scores'][0], 2) >= min_score:
        categorized_sscorpus_full.append([ simple_list[index], result['sequence'], result['labels'][0], round(result['scores'][0], 2) ])

with open(dump_root_path+'initial/sscorpus_categorized_preprocessed_full.csv', 'w+', newline='') as f:
    write = csv.writer(f)
    write.writerows(categorized_sscorpus_full)
    f.close()

# Convert list into pandas Dataframe for easier analysis

#categorized_columns = ["Simple", "Professional", "Category", "Score"]
#df_sscorpus_categorized = pd.DataFrame(categorized_sscorpus_specific, columns=categorized_columns)

# Analysis for the spreading of entries according to category

#category = "Sports"
#df_sscorpus_categorized[df_sscorpus_categorized.Category == category].shape[0]
