==> 3.1_extract_sample_datasets_multi_label_true.ipynb <==
import pandas as pd

# Load categorized corpus when multi_label is True

columns = ['Simple', 'Professional', 'Category', 'Score']
sscorpus_categorized_path = "2_sscorpus_categorized/multi_label_true/initial/sscorpus_categorized_preprocessed_specific.csv"
df_sscorpus_categorized = pd.read_csv(sscorpus_categorized_path, sep=',', header=None, names=columns)
#df_sscorpus_categorized.head(5)
df_sscorpus_categorized.shape

# Extract 100 samples for every category and write into one CSV file

df_final_sample_dataset = pd.DataFrame(columns=columns)
categories = ["Business", "Entertainment", "Geography", "Politics", "Science", "Sports"]
annotations_root_path = "3_sscorpus_sample_datasets_annotated/multi_label_true/final/"
for category in categories:
    df_category = df_sscorpus_categorized.loc[(df_sscorpus_categorized['Category'] == category)].sample(n=100)
    df_final_sample_dataset = df_final_sample_dataset.append(df_category, ignore_index=True)

df_final_sample_dataset.to_csv(annotations_root_path+"annotated_sample.csv", sep=',', header=None, index=False, encoding='utf-8')
