==> 4.0_analyze_annotations_multi_label_false.ipynb <==
import pandas as pd

# Load the annotated dataset

# set the path of the annotated dataset
path = "3_sscorpus_sample_datasets_annotated/multi_label_false/final/annotated_Business.csv"

columns = ['Simple', 'Professional', 'Category', 'Score', 'Annotation']
df_sscorpus_annotated = pd.read_csv(path, sep=',', header=None, names=columns)
#df_sscorpus_annotated.head(5)
df_sscorpus_annotated.shape
#df_sscorpus_annotated.dtypes

# Analysis to make sure that no entry is forgotten to be annotated

#df_sscorpus_annotated['Annotation'].isnull().sum()
df_sscorpus_annotated[df_sscorpus_annotated['Annotation'].isnull()]

# Set the minimum probability score

score = 0.38

# Annotated samples that are part of the respective category and have a probability higher than the min score

annotated_samples = df_sscorpus_annotated.loc[(df_sscorpus_annotated['Annotation'] == 1) & (df_sscorpus_annotated['Score'] >= score)].shape[0]

# All samples that have a probability higher than the min score

all_samples = df_sscorpus_annotated.loc[(df_sscorpus_annotated['Score'] >= score)].shape[0]

# Try out different minimum probability scores and stop when reaching at least 90 percent

print(annotated_samples)
print(all_samples)
print(annotated_samples/all_samples)
