==> 5.0_build_final_datasets_multi_label_false.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

# Load the corpus for multi_label as False

columns = ['Simple', 'Professional', 'Category', 'Score']
path = "2_sscorpus_categorized/multi_label_false/scorpus_categorized_raw"
df_sscorpus_categorized = pd.read_csv(path, sep=',', header=None, names=columns)
#df_sscorpus_categorized.head(5)
df_sscorpus_categorized.shape

# Set the minimum probability scores for each category that were obtained after annotations

datasets_info = [ ["Business", 0.38], ["Entertainment", 0.37], ["Geography", 0.3], 
                  ["Politics", 0.39], ["Science", 0.3], ["Sports", 0.27] ]

# Create datasets based on category and split them into train, validation, test

for dataset_info in datasets_info:
    category = dataset_info[0]
    score = dataset_info[1]
    df_dataset = df_sscorpus_categorized.loc[(df_sscorpus_categorized['Category'] == category) & (df_sscorpus_categorized['Score'] >= score)]
    df_dataset_subset = df_dataset[['Simple', 'Professional']]
    df_dataset_subset.to_csv("4_sscorpus_final_datasets/multi_label_false/"+category+"/"+category.lower()+".csv", sep=',', header=None, index=False, encoding='utf-8')
    df_train, df_tmp = train_test_split(df_dataset_subset, test_size=0.2)
    df_eval, df_test = train_test_split(df_tmp, test_size=0.5)
    df_train.to_csv("4_sscorpus_final_datasets/multi_label_false/"+category+"/train/train.csv", sep=',', header=None, index=False, encoding='utf-8')
    df_eval.to_csv("4_sscorpus_final_datasets/multi_label_false/"+category+"/eval/eval.csv", sep=',', header=None, index=False, encoding='utf-8')
    df_test.to_csv("4_sscorpus_final_datasets/multi_label_false/"+category+"/test/test.csv", sep=',', header=None, index=False, encoding='utf-8')

# Create the corpus based on concatenation of all 6 datasets and split it into train, validation, test

df_business = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Business/business.csv", sep=',', header=None, names=['Simple', 'Professional'])
df_entertainment = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Entertainment/entertainment.csv", sep=',', header=None, names=['Simple', 'Professional'])
df_geography = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Geography/geography.csv", sep=',', header=None, names=['Simple', 'Professional'])
df_politics = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Politics/politics.csv", sep=',', header=None, names=['Simple', 'Professional'])
df_science = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Science/science.csv", sep=',', header=None, names=['Simple', 'Professional'])
df_sports = pd.read_csv("4_sscorpus_final_datasets/multi_label_false/Sports/sports.csv", sep=',', header=None, names=['Simple', 'Professional'])

frames = [df_business, df_entertainment, df_geography, df_politics, df_science, df_sports]
df_all = pd.concat(frames)
df_all.to_csv("4_sscorpus_final_datasets/multi_label_false/All/all.csv", sep=',', header=None, index=False, encoding='utf-8')
df_all_train, df_all_tmp = train_test_split(df_all, test_size=0.2)
df_all_eval, df_all_test = train_test_split(df_all_tmp, test_size=0.5)
df_all_train.to_csv("4_sscorpus_final_datasets/multi_label_false/All/train/train.csv", sep=',', header=None, index=False, encoding='utf-8')
df_all_eval.to_csv("4_sscorpus_final_datasets/multi_label_false/All/eval/eval.csv", sep=',', header=None, index=False, encoding='utf-8')
df_all_test.to_csv("4_sscorpus_final_datasets/multi_label_false/All/test/test.csv", sep=',', header=None, index=False, encoding='utf-8')
