==> 5.1_build_final_datasets_multi_label_true.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

# Load the corpus for multi_label as True

columns = ['Simple', 'Professional', 'Category', 'Score']
path_specific = "2_sscorpus_categorized/multi_label_true/sscorpus_categorized_preprocessed_specific.csv"
df_sscorpus_categorized = pd.read_csv(path_specific, sep=',', header=None, names=columns)
#df_sscorpus_categorized.head(5)
df_sscorpus_categorized.shape

# Set the 6 categories

categories = ["Business", "Entertainment", "Geography", "Politics", "Science", "Sports"]

for category in categories:
    df_dataset = df_sscorpus_categorized.loc[(df_sscorpus_categorized['Category'] == category)]
    df_dataset_subset = df_dataset[['Simple', 'Professional']]
    df_dataset_subset.to_csv("4_sscorpus_final_datasets/multi_label_true/"+category+"/"+category.lower()+".csv", sep=',', header=None, index=False, encoding='utf-8')
    df_train, df_tmp = train_test_split(df_dataset_subset, test_size=0.2)
    df_eval, df_test = train_test_split(df_tmp, test_size=0.5)
    df_train.to_csv("4_sscorpus_final_datasets/multi_label_true/"+category+"/train/train.csv", sep=',', header=None, index=False, encoding='utf-8')
    df_eval.to_csv("4_sscorpus_final_datasets/multi_label_true/"+category+"/eval/eval.csv", sep=',', header=None, index=False, encoding='utf-8')
    df_test.to_csv("4_sscorpus_final_datasets/multi_label_true/"+category+"/test/test.csv", sep=',', header=None, index=False, encoding='utf-8')

# Create the corpus and split it into train, validation, test

columns = ['Simple', 'Professional', 'Category', 'Score']
path_full = "2_sscorpus_categorized/multi_label_true/sscorpus_categorized_preprocessed_full.csv"
df_sscorpus_all = pd.read_csv(path_full, sep=',', header=None, names=columns)
#df_sscorpus_all.head(5)
df_sscorpus_all.shape

df_all = df_sscorpus_all[['Simple', 'Professional']]

df_all.to_csv("4_sscorpus_final_datasets/multi_label_true/All/all.csv", sep=',', header=None, index=False, encoding='utf-8')

df_all_train, df_all_tmp = train_test_split(df_all, test_size=0.2)
df_all_eval, df_all_test = train_test_split(df_all_tmp, test_size=0.5)
df_all_train.to_csv("4_sscorpus_final_datasets/multi_label_true/All/train/train.csv", sep=',', header=None, index=False, encoding='utf-8')
df_all_eval.to_csv("4_sscorpus_final_datasets/multi_label_true/All/eval/eval.csv", sep=',', header=None, index=False, encoding='utf-8')
df_all_test.to_csv("4_sscorpus_final_datasets/multi_label_true/All/test/test.csv", sep=',', header=None, index=False, encoding='utf-8')
