==> 7.0_bart_predictions.ipynb <==
import logging
import pandas as pd
from simpletransformers.seq2seq import Seq2SeqModel, Seq2SeqArgs

logging.basicConfig(level=logging.INFO)
transformers_logger = logging.getLogger("transformers")
transformers_logger.setLevel(logging.WARNING)

# Load the test dataset

# choose the dataset
dataset = "Sports"
columns = ["Simple", "Professional"]
#set the desired path
path = "4_sscorpus_final_datasets/multi_label_true/"+dataset+"/test/test.csv"
df_test = pd.read_csv(path, sep=',', header=None, names=columns)
#df_test.head(5)
df_test.shape

# Load the finetuned model

# choose the finetuned model
finetuned_model = "All"

model_path = "outputs/multi_label_true/"+finetuned_model+"/"

model = Seq2SeqModel(
    encoder_decoder_type="bart",
    encoder_decoder_name=model_path,
    use_cuda=True
)

predictions = model.predict(df_test['Simple'].values.tolist())

# Perform predictions

df_predictions = pd.DataFrame(predictions, columns = ['Prediction'])

df_predictions.shape

# Write generated predictions and reference ground truths together in a Dataframe

df_predictions_final = pd.concat([df_test, df_predictions], axis=1)

df_predictions_final.shape

# Write the generated and reference sentences into a file

predictions_path = "5_predictions/multi_label_true/"+dataset+"/predictions_"+dataset+"_dataset_with_"+finetuned_model+"_model.csv"

df_predictions_final.to_csv(predictions_path, sep=',', header=None, index=False, encoding='utf-8')
