==> 8.0_calculate_metrics.ipynb <==
import pandas as pd

# Set the path for the baseline predictions (corpus) and the one we want to compare (specific dataset) 

columns = ["Source", "Target", "Prediction"]
dataset = "Sports"

path_specific_model = "5_predictions/multi_label_true/"+dataset+"/predictions_"+dataset+"_dataset_with_"+dataset+"_model.csv"
path_full_model = "5_predictions/multi_label_true/"+dataset+"/predictions_"+dataset+"_dataset_with_All_model.csv"

# Load the predictions into Dataframes

df_prediction_specific = pd.read_csv(path_specific_model, sep=',', header=None, names=columns)
df_prediction_full = pd.read_csv(path_full_model, sep=',', header=None, names=columns)

print(df_prediction_specific.shape)
print(df_prediction_full.shape)

# Calculate ROUGE metrics

from rouge import Rouge

rouge = Rouge()
rouge_scores_specific = rouge.get_scores(df_prediction_specific["Prediction"], df_prediction_specific["Target"], avg=True)
rouge_scores_full = rouge.get_scores(df_prediction_full["Prediction"], df_prediction_full["Target"], avg=True)

import pprint

print(f'\n\nROUGE scores for {dataset} dataset with {dataset} model. \n')
pprint.pprint(rouge_scores_specific)

print(f'\n\nROUGE scores for {dataset} dataset with All model. \n')
pprint.pprint(rouge_scores_full)

# Calculate BLEU metrics

from nltk.translate.bleu_score import sentence_bleu

def calculate_bleu(predictions, targets):
    score_bleu1 = 0.
    score_bleu2 = 0.

    for i in range(len(predictions)):
        score_bleu1 += sentence_bleu([targets[i].split()], predictions[i].split(), weights=(1, 0, 0, 0))
    score_bleu1 /= len(predictions)
    
    for i in range(len(predictions)):
        score_bleu2 += sentence_bleu([targets[i].split()], predictions[i].split(), weights=(0, 1, 0, 0))
    score_bleu2 /= len(predictions)
    
    return [score_bleu1, score_bleu2]

bleu_scores_specific = calculate_bleu(df_prediction_specific["Prediction"], df_prediction_specific["Target"])
bleu_scores_full = calculate_bleu(df_prediction_full["Prediction"], df_prediction_full["Target"])

print(f'\n\BLEU scores for {dataset} dataset with {dataset} model. \n')
pprint.pprint(bleu_scores_specific)

print(f'\n\BLEU scores for {dataset} dataset with All model. \n')
pprint.pprint(bleu_scores_full)

# Calculate BERTScore metrics

from evaluate import load
from statistics import mean

bertscore = load("bertscore")
model = "microsoft/deberta-xlarge-mnli"

predictions_specific = df_prediction_specific["Prediction"]
references_specific = df_prediction_specific["Target"]
results_specific = bertscore.compute(predictions=predictions_specific, references=references_specific, lang="en", model_type=model)
#pprint.pprint(results_specific)
f1_score_specific = mean(results_specific["f1"])
precision_specific = mean(results_specific["precision"])
recall_specific = mean(results_specific["recall"])

print(f'BERTScore for {dataset} dataset with {dataset} model. \n')
print(f'F1: {f1_score_specific}')
print(f'Precision: {precision_specific}')
print(f'Recall: {recall_specific}')

predictions_full = df_prediction_full["Prediction"]
references_full = df_prediction_full["Target"]
results_full = bertscore.compute(predictions=predictions_full, references=references_full, lang="en", model_type=model)
#pprint.pprint(results_full)
f1_score_specific = mean(results_full["f1"])
precision_specific = mean(results_full["precision"])
recall_specific = mean(results_full["recall"])

print(f'BERTScore for {dataset} dataset with All model. \n')
print(f'F1: {f1_score_specific}')
print(f'Precision: {precision_specific}')
print(f'Recall: {recall_specific}')
